Demonstrating Touch Events
Attempting to get touch events working on the PiStorm V2

the major changes are as follows:

ConsoleApp.vala: Added a second entry in 'event' to detect touch events
Reading in all touch events and sending the full touch event to be
enqueued

Screen.vala:
* Added a touch queue to absorb the asynchronous touch events
* Added functions which will find widgets at touch coordinates
* Modified 'handle_input' to process touch events

Button.vala:
* As a test press the button when the user presses the appropriate
part of the screen, specifically when the user touches the button
the button should be 'pressed'

Widget.vala: Added a new virtual touch event to absorb any unhandled
touch events

Notes:

The touch events are not behaving as I anticipated, the following
issues were found.

1. TOUCH_UP is never detected when the user raises the touch input
I believe the touch event should be '12' which should print out
on the console as a critical message.
2. The screen is uncalibrated on the bottom half and roughly 2/3rds
of the screen registers inputs and these values are somehow mapped
to the rest of the screen. This makes me think there is something
wrong with the 'rotation' settings